new file testProxyRestore
